Optimized check_enough_disk_space and adding option to run main() automatically from command line
Oprtimized check_enough_disk_space separating the calculation of the size of the home folder with the one of its subfolders
because these take much longer. So the subfolder sizes are only calculated if the check doesn't pass (not enough space)

Also used sys.argv so that if the user adds the "auto": string when running cpu_health_checks it automatically runs the main()
function with no extra arguments, to avoid problems for users that don't have the interactive option of python.